Added mcu firmware for rev1 of board
This codebase was intended for r2 of the primecam amp frontend board
however, rev1 is now supported as well. To prevent a migraine, the bossa
flash utility used by arduino has been compiled from
https://github.com/shumatech/BOSSA/tree/arduino
and commited here. Update code may not be needed for r2 and thats why
only the prototype bin files are present.